Create devcontainer_utils to handle most of the heavy work
